+++ src/main/java/com/amkor/service/ATVService.java
@@ -319,128 +319,6 @@ public class ATVService implements IATVService {
     }
 
     @Override
-    public boolean checkExistAutoLabel(AutoLabelModel model) {
-        boolean result = false;
-        Connection m_conn;
-        PreparedStatement m_psmt;
-        ResultSet m_rs;
-        try {
-            String sQuery = "select * from EMLIB.EAUTOLBLVP where FACTORY_ID = ? AND SITE_ID = ? AND BUSINESS_TYPE = ? " +
-                    "AND CUSTOMER = ? AND PACKAGE = ? AND DIMENSION = ? AND LEAD = ? AND TARGET_DEVICE = ? " +
-                    "AND KEY_FIELD1 = ? AND KEY_FIELD2 = ? AND FIELD_NAME = ?";
-            Class.forName(this.getDriver());
-            m_conn = DriverManager.getConnection(getURL(SharedConstValue.AMKOR_SHORTNAME), getUserID(SharedConstValue.AMKOR_SHORTNAME), getPasswd(SharedConstValue.AMKOR_SHORTNAME));
-            m_psmt = m_conn.prepareStatement(sQuery);
-            int i = 1;
-            m_psmt.setInt(i++, model.getFactoryId());
-            m_psmt.setInt(i++, model.getSiteId());
-            m_psmt.setString(i++, model.getBusinessType());
-            m_psmt.setInt(i++, model.getCustomerId());
-            m_psmt.setString(i++, model.getPkg());
-            m_psmt.setString(i++, model.getDim());
-            m_psmt.setString(i++, model.getLead());
-            m_psmt.setString(i++, model.getTargetDevice());
-            m_psmt.setString(i++, model.getKeyField1());
-            m_psmt.setString(i++, model.getKeyField2());
-            m_psmt.setString(i++, model.getFieldName());
-            m_rs = m_psmt.executeQuery();
-            if (m_rs.next()) {
-                result = true;
-            }
-
-            m_conn.close();
-            m_psmt.close();
-            m_rs.close();
-
-        } catch (Exception ex) {
-            log.error(ex.getMessage(), ex);
-        }
-
-
-        return result;
-    }
-
-    @Override
-    public OnLineScheduleSheetFileModel getOnlineScheduleSheetMemoFileFromStationAndLotName(String station, String lotName) {
-        Connection m_conn;
-        PreparedStatement m_psmt;
-        ResultSet m_rs;
-        OnLineScheduleSheetFileModel onLineScheduleSheetFileModel = null;
-        try {
-            String sQuery = "SELECT * FROM EMLIB.EMESTP032 " +
-                    "join EMLIB.ASCHMP02 on FACTORY_ID = SMFCID AND RECORD_ID = SMSCH# " +
-                    "WHERE FACTORY_ID = " + SharedConstValue.FACTORY_ID + " AND TYPE_ID = 'S' AND SMLOT# = '" + lotName + "' AND ( " +
-                    "EXISTS (SELECT 1 FROM EMLIB.EMESTP032 WHERE SMLOT# = '" + lotName + "' AND FILE_X LIKE '%" + station + "%') " +
-                    "OR FILE_X LIKE '%" + station + "%')";
-            Class.forName(this.getDriver());
-            m_conn = DriverManager.getConnection(getURL(SharedConstValue.AMKOR_SHORTNAME), getUserID(SharedConstValue.AMKOR_SHORTNAME), getPasswd(SharedConstValue.AMKOR_SHORTNAME));
-            m_psmt = m_conn.prepareStatement(sQuery);
-            m_rs = m_psmt.executeQuery();
-            if (m_rs.next()) {
-                onLineScheduleSheetFileModel = new OnLineScheduleSheetFileModel();
-
-                onLineScheduleSheetFileModel.setFactoryID(m_rs.getInt(1));
-                onLineScheduleSheetFileModel.setType(m_rs.getString(2));
-                onLineScheduleSheetFileModel.setRecordID(m_rs.getString(3));
-                onLineScheduleSheetFileModel.setPath(m_rs.getString(4));
-                onLineScheduleSheetFileModel.setFile(m_rs.getString(5));
-                onLineScheduleSheetFileModel.setEffectiveTo(m_rs.getLong(6));
-                onLineScheduleSheetFileModel.setCreateDateTime(m_rs.getLong(7));
-                onLineScheduleSheetFileModel.setCreateBadge(m_rs.getString(8));
-                onLineScheduleSheetFileModel.setMaintDateTime(m_rs.getLong(9));
-                onLineScheduleSheetFileModel.setMaintBadge(m_rs.getString(10));
-            }
-
-            m_conn.close();
-            m_psmt.close();
-            m_rs.close();
-        } catch (Exception ex) {
-            log.error(ex.getMessage(), ex);
-        }
-
-        return onLineScheduleSheetFileModel;
-    }
-
-    @Override
-    public boolean checkExistProcessNote(ProcessNoteModel model) {
-
-        boolean result = false;
-        Connection m_conn;
-        PreparedStatement m_psmt;
-        ResultSet m_rs;
-        try {
-            String sQuery = "select * from EPLIB.EPENOTP where ENFCID = ? AND ENCLAS = ? AND ENCUST = ? AND ENPKGE = ? " +
-                    "AND ENDMSN = ? AND ENLEAD = ? AND ENDEVC = ? AND ENOPID = ? AND ENOPER = ? AND ENSEQ# = ?";
-            Class.forName(this.getDriver());
-            m_conn = DriverManager.getConnection(getURL(SharedConstValue.AMKOR_SHORTNAME), getUserID(SharedConstValue.AMKOR_SHORTNAME), getPasswd(SharedConstValue.AMKOR_SHORTNAME));
-            m_psmt = m_conn.prepareStatement(sQuery);
-            int i = 1;
-            m_psmt.setInt(i++, model.getFactoryId());
-            m_psmt.setString(i++, model.getClassify());
-            m_psmt.setInt(i++, model.getCustomerId());
-            m_psmt.setString(i++, model.getPkg());
-            m_psmt.setString(i++, model.getDim());
-            m_psmt.setString(i++, model.getLead());
-            m_psmt.setString(i++, model.getTargetDevice());
-            m_psmt.setString(i++, model.getOptionId());
-            m_psmt.setInt(i++, model.getOperation());
-            m_psmt.setInt(i, model.getSeq());
-            m_rs = m_psmt.executeQuery();
-            if (m_rs.next()) {
-                result = true;
-            }
-
-            m_rs.close();
-            m_psmt.close();
-            m_conn.close();
-        } catch (Exception ex) {
-            log.error(ex.getMessage(), ex);
-        }
-
-        return result;
-    }
-
-    @Override
     public List<AlertForFGModel> getAlertForFGNotScheduledFor30Days(int factoryId, String plant, String cust) {
         Connection m_conn;
         PreparedStatement m_psmt;
