+++ src/main/java/com/amkor/service/iService/IATVService.java
@@ -11,10 +11,4 @@ import java.util.List;
 @Service
 public interface IATVService extends IReadService {
     List<AlertForFGModel> getAlertForFGNotScheduledFor30Days(int factoryId, String plant, String cust);
-
-    boolean checkExistProcessNote(ProcessNoteModel model);
-
-    boolean checkExistAutoLabel(AutoLabelModel model);
-
-    OnLineScheduleSheetFileModel getOnlineScheduleSheetMemoFileFromStationAndLotName(String station, String lotName);
 }
