+++ src/main/java/com/amkor/service/iService/ITFAService.java
@@ -2,13 +2,14 @@ package com.amkor.service.iService;
 
 import com.amkor.models.ApiLoggingModel;
 import com.amkor.models.AutoLabelModel;
+import com.amkor.models.OnLineScheduleSheetFileModel;
 import com.amkor.models.ProcessNoteModel;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public interface IATVThanhService extends IWriteService {
+public interface ITFAService extends IWriteService {
     boolean sendMailProcess(String title, String sContent, List<String> toPeople, List<String> ccPeople, List<String> fileNames);
 
     int updateProcessNote(ProcessNoteModel model);
@@ -19,5 +20,10 @@ public interface IATVThanhService extends IWriteService {
 
     int createProcessNote(ProcessNoteModel model);
 
-    int addApiLogging(ApiLoggingModel model);
+    boolean checkExistProcessNote(ProcessNoteModel model);
+
+    boolean checkExistAutoLabel(AutoLabelModel model);
+
+    OnLineScheduleSheetFileModel getOnlineScheduleSheetMemoFileFromStationAndLotName(String station, String lotName);
+
 }
